Implement preval.
