+++ package.json
@@ -6,9 +6,11 @@
   "license": "MIT",
   "repository": "Qix-/color-convert",
   "scripts": {
+    "build": "rollup -c",
     "pretest": "xo",
-    "test": "node test/basic.js"
+    "test": "npm run build && node test/basic.js"
   },
+  "main": "lib/index.js",
   "engines": {
     "node": ">=7.0.0"
   },
@@ -27,10 +29,7 @@
     "ansi16"
   ],
   "files": [
-    "index.js",
-    "conversions.js",
-    "css-keywords.js",
-    "route.js"
+    "lib"
   ],
   "xo": {
     "rules": {
@@ -40,7 +39,13 @@
     }
   },
   "devDependencies": {
+    "@babel/core": "^7.4.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.4.0",
+    "@babel/register": "^7.4.0",
+    "babel-plugin-codegen": "^3.0.0",
     "chalk": "^2.4.2",
+    "rollup": "^1.7.4",
+    "rollup-plugin-babel": "^4.3.2",
     "xo": "^0.24.0"
   },
   "dependencies": {
