+++ src/index.js
@@ -1,0 +1,1 @@
+codegen.require('./models'); // eslint-disable-line no-undef
